#31 Moves to VS 2017 platform toolset (v141)

+++ TestShell/Simulations/Particles/RandomWalk.cpp
@@ -58,7 +58,7 @@ RandomWalk::~RandomWalk() {}
 
 void RandomWalk::Init() {
     for (auto& p : particles) p.Die();
-    iota(begin(indicesDead), end(indicesDead), 0);
+    iota(begin(indicesDead), end(indicesDead), (uint16)0);
     deadCount = kParticlesCount;
     matParticle_ = make_unique<MaterialRaii>(viewport_.CreateMaterial(MaterialType::kWhite(), "rigid"));
